requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
tensorflow>=2.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
joblib
